refactor: DB 필드명 수정
