--- src/main/java/chess/dao/GameDao.java
@@ -1,85 +1,0 @@
-package chess.dao;
-
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.sql.Statement;
-
-import chess.domain.game.ChessGame;
-
-public class GameDao {
-
-    private static final String URL = "jdbc:mysql://localhost:3306/chess";
-    private static final String USER = "username";
-    private static final String PASSWORD = "password";
-
-    private final Connection connection = getConnection();
-
-    public int save(ChessGame chessGame) {
-        String sql = "INSERT INTO game (state) values (?)";
-
-        try (PreparedStatement statement = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
-            statement.setString(1, chessGame.getState().getState());
-            statement.executeUpdate();
-            ResultSet resultSet = statement.getGeneratedKeys();
-            if (!resultSet.next()) {
-                throw new SQLException();
-            }
-            return resultSet.getInt(1);
-        } catch (SQLException e) {
-            e.printStackTrace();
-            throw new IllegalArgumentException();
-        }
-    }
-
-    public String findState(String id) {
-        String sql = "SELECT state FROM game WHERE id = ?";
-
-        try (PreparedStatement statement = connection.prepareStatement(sql)) {
-            statement.setString(1, id);
-            ResultSet resultSet = statement.executeQuery();
-            if (!resultSet.next()) {
-                throw new SQLException("쿼리문 실행 결과가 존재하지 않습니다.");
-            }
-            return resultSet.getString("state");
-        } catch (SQLException e) {
-            e.printStackTrace();
-            throw new IllegalArgumentException();
-        }
-    }
-
-    public void updateById(String id, String state) {
-        String sql = "UPDATE game SET state = ? where id = ?";
-
-        try (PreparedStatement statement = connection.prepareStatement(sql)) {
-            statement.setString(1, state);
-            statement.setString(2, id);
-            statement.executeUpdate();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-    }
-
-    public void delete(String id) {
-        String sql = "DELETE FROM game WHERE id = ?";
-
-        try (PreparedStatement statement = connection.prepareStatement(sql)) {
-            statement.setString(1, id);
-            statement.executeUpdate();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-    }
-
-    private Connection getConnection() {
-        Connection connection = null;
-        try {
-            connection = DriverManager.getConnection(URL, USER, PASSWORD);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return connection;
-    }
-}
